id da Long a Integer

+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/application/dto/PianoDTO.java
@@ -17,6 +17,6 @@ public class PianoDTO implements Serializable {
 
     private String nome;
     private int numero;
-    private Long sede;
+    private Integer sede;
 
 }
