+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/application/dto/StanzaDTO.java
@@ -18,7 +18,7 @@ public class StanzaDTO implements Serializable {
 
     private String nome;
     private int capienza;
-    private Long piano;
-    private List<Long> posti;
+    private Integer piano;
+    private List<Integer> posti;
 
 }
