+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/application/usecases/spec/ApprovaPrenotazione.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Component;
 
 @Component
 public interface ApprovaPrenotazione {
-    public PrenotazioneDTO approvaPrenotazione(Long id);
+    public PrenotazioneDTO approvaPrenotazione(Integer id);
 }
