+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/domain/repository/PostoRepository.java
@@ -3,5 +3,5 @@ package com.majorbit.bozza_proj_turni_prenotazioni.domain.repository;
 import com.majorbit.bozza_proj_turni_prenotazioni.domain.model.Posto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostoRepository  extends JpaRepository<Posto, Long> {
+public interface PostoRepository  extends JpaRepository<Posto, Integer> {
 }
